Wrapp FileService in FileModule
