+++ geoprocessing/src/modules/files/files.module.ts
@@ -1,0 +1,8 @@
+import { FileService } from './files.service';
+import { Logger, Module } from '@nestjs/common';
+
+@Module({
+  providers: [FileService, Logger],
+  exports: [FileService],
+})
+export class FilesModule {}
