+++ geoprocessing/src/modules/planning-units/planning-units.module.ts
@@ -7,7 +7,7 @@ import { PlanningUnitsController } from './planning-units.controller';
 import { ShapefileService } from '../shapefiles/shapefiles.service';
 import { PlanningUnitsGeom } from 'src/modules/planning-units/planning-units.geo.entity';
 import { PlanningUnitsService } from './planning-units.service';
-import { FileService } from 'src/utils/file.utils';
+import { FileService } from '../files/files.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([PlanningUnitsGeom]), TileModule],
