+++ geoprocessing/src/utils/file.utils.ts
@@ -6,10 +6,6 @@ import { MulterOptions } from '@nestjs/platform-express/multer/interfaces/multer
 import { Request } from 'express';
 import { diskStorage } from 'multer';
 import * as tempDirectory from 'temp-dir';
-import { createReadStream } from 'fs';
-import { unlink, rmdir } from 'fs/promises';
-import { Extract } from 'unzipper';
-import * as path from 'path';
 
 export const uploadOptions: Readonly<MulterOptions> = {
   storage: diskStorage({
@@ -29,36 +25,3 @@ export const uploadOptions: Readonly<MulterOptions> = {
     },
   }),
 };
-
-export class FileService {
-  unzipFile(fileInfo: Express.Multer.File): Promise<string> {
-    return new Promise((resolve, reject) => {
-      createReadStream(fileInfo.path)
-        .pipe(
-          Extract({
-            path: path.join(
-              fileInfo.destination,
-              path.basename(fileInfo.filename.replace('.zip', '')),
-            ),
-          }),
-        )
-        .on('close', () =>
-          resolve(`${fileInfo.filename} extracted successfully`),
-        )
-        .on('error', (error: Error) =>
-          reject(
-            new Error(`${fileInfo.filename} could not be extracted: ` + error),
-          ),
-        );
-    });
-  }
-
-  deleteDataFromFS(path: string): void {
-    if (path.startsWith('/tmp')) {
-      unlink(path);
-      rmdir(path.replace('.zip', ''), { recursive: true });
-    } else {
-      throw new Error(`Could not complete deletion: ${path} is not in /tmp`);
-    }
-  }
-}
